:sparkles: New `Universal.ControlStructures.IfElseDeclaration` sniff
New sniff to verify that `else(if)` statements with braces are on a new line.

Sister-sniff to the following two PHPCS native sniffs which each demand that `else[]if` is on the same line as the closing curly of the preceding `(else)if`:
- `PEAR.ControlStructures.ControlSignature[.Found]`
- `Squiz.ControlStructures.ControlSignature.SpaceAfterCloseBrace`

Other related sniffs:
- `Squiz.ControlStructures.ElseIfDeclaration` Forbids the use of "elseif", demands "else if".
- `PSR2.ControlStructures.ElseIfDeclaration`  Forbids the use of "else if", demands "elseif".

Includes fixer.
Includes unit tests.
Includes documentation.
Includes metrics.