:sparkles: New `NormalizedArrays.Arrays.CommaAfterLast` sniff
Configurable sniff to enforce/forbid a comma after the last array item.

By default, this sniff will:
- forbid a comma after the last array item for _single-line arrays_.
- enforce a comma after the last array item for _multi-line arrays_.

This can be changed for each type or array individually by setting the `singleLine` or `multiLine` properties in a custom ruleset.
The valid values are: `enforce`, `forbid` or `skip` to not check the comma after the last array item for a particular type of array.

Includes fixers.
Includes unit tests.
Includes documentation.
Includes metrics.