Composer: add PHPCSDevCS dependency
The `PHPCSDev` ruleset has been split off from PHPCSDevTools to its own package.

The new `PHPCSDevCS` package has a minimum PHPCS requirement of PHPCS 3.5.0 (this was previously also the case, but not enforced).
To prevent issues with Composer refusing to install a PHPCS version below 3.5.0, the dependency has not been added to `require-dev`, but instead is pulled in via a `composer require` in a composer script.

Also: `PHPCSDevTools` 1.0.0 has been released :tada:

Refs:
* https://github.com/PHPCSStandards/PHPCSDevCS
* PHPCSStandards/PHPCSDevTools 29