:sparkles: New `Universal.UseStatements.NoLeadingBackslash` sniff
New sniff to verify that a name being imported in an import use statement does not start with a leading backslash.

Names in import `use` statements should always be fully qualified, so a leading backslash is not needed and it is strongly recommended not to use one.

This sniff handles all types of import use statements supported by PHP, in contrast to any of the other sniffs for the same in, for instance, the PSR12 or the Slevomat standard, all of which are incomplete.

Includes fixer.
Includes unit tests.
Includes documentation.
Includes metrics.