:sparkles: New `Universal.Namespaces.DisallowDeclarationWithoutName` sniff
New sniff to disallow the use of namespace declarations without a namespace name.

Declaring a namespace declaration without a namespace name is only possible using the curly brace syntax and makes it so the code within the braces is considered to be in the global namespace.

A namespace declaration without curly braces without a name would be a parse error and will be ignored for the purposes of this sniff.

Includes unit tests.
Includes documentation.
Includes metrics.