:sparkles: New `Universal.Operators.DisallowLogicalAndOr` sniff
New sniff to enforce the use of the boolean `&&` and `||` operators instead of the logical `and`/`or` operators.

Note: as the [operator precedence](https://www.php.net/manual/en/language.operators.precedence.php) of the logical operators is significantly lower than the operator precedence of boolean operators, this sniff does not contain an auto-fixer.

Includes unit tests.
Includes documentation.
Includes metrics.