Universal/DuplicateArrayKey: add support for detecting duplicate key PHP cross-version
How array keys for array items without keys are being determined, has changed in PHP 8.0.

Previously, the key would be the highest previously seen numeric key + 1, providing the highest previously seen numeric key was 0 or higher. Otherwise, it would be 0.

As of PHP 8.0, the key will be the highest previously seen numeric key + 1, independently of whether the previously seen numeric key was negative.

The sniff will now calculate and track the keys for unkeyed array items using both the PHP < 8.0 logic as well as the PHP >= 8.0 logic.

If a duplicate key would yield the same error in both PHP < 8.0 as well as PHP >= 8.0, the `Found` error code will be used.
If a duplicate key would only be a duplicate on PHP < 8.0, the `FoundForPHPlt80` error code will be used and the error message will indicate the error only applies to PHP < 8.0.
If a duplicate key would only be a duplicate on PHP > 8.0, the `FoundForPHPgte80` and the error message will indicate the error only applies to PHP >= 8.0.

If the end-user has set the `php_version` configuration option, the sniff will respect that and only report duplicate keys for the PHP version indicated.

Includes additional unit tests.

P.S.: the code could possible be made smarter, but for now, this is fine as the goal of accounting for this change in PHP has been achieved.

Refs:
* https://3v4l.org/liBXD
* https://wiki.php.net/rfc/negative_array_index
* https://github.com/squizlabs/PHP_CodeSniffer/wiki/Configuration-Options#setting-the-php-version