QA/CI: change the way the CS check is run
Use `composer global` instead of a project local install for the CS check as PHPCSExtra will become a dependency of PHPCSDevCS and it would otherwise create a conflicting, circular dependency.

To make this simpler, split the old "Sniff" stage into "Sniff" and "Validate" and skip the `install` and `before_script` steps for the "Sniff" stage.