Tests: Xdebug 3 compatibility fix
As the tests are based on the PHPCS native test framework, they are limited to max PHPUnit 7.5.x.

However, Xdebug `3.0` was released last week and the Travis images for high PHP versions have been updated to include Xdebug `3.0`, but, of course, PHPUnit 7.5 is not compatible with Xdebug 3....

Declaring the "missing" constants which no longer exist in Xdebug 3, but which PHPUnit 7.5 is relying on to exist, will fix this, combined with enabling the Xdebug code coverage mode from within the Travis config.