:sparkles: New `Universal.ControlStructures.DisallowLonelyIf` sniff
Sniff to disallow `if` statements as the only statement in an `else` block.

Inspired by the [ESLint "no lonely if"](https://eslint.org/docs/rules/no-lonely-if) rule in response to upstream PHPCS issue 3206.

The sniff, as well as the fixer, will handle both curly brace if/else control structures as well as alternative syntax - if/else/endif - control structures correctly.
Inline control structures, i.e. without curly braces and not using alternative syntax, will be ignored by the sniff.

Note: This sniff will not fix the indentation of the "inner" code. It is strongly recommended to run this sniff together with the `Generic.WhiteSpace.ScopeIndent` sniff to get the correct indentation.

Includes fixer.
Includes unit tests.
Includes documentation.