GH Actions: fix downloading package from source
The test run against PHPCS 4.x needs to download from `source` instead of Packagist to ensure the necessary test base classes are available.

Most of the time, this fails with an error:
```
   [Composer\Downloader\TransportException]
   Could not authenticate against github.com
```

Composer then downloads from Packagist anyway which will cause the tests to fail as the test runners aren't available.

Passing the Composer Token to the `setup-php` action should fix this.

Refs:
* https://github.com/shivammathur/setup-php#composer-github-oauth
* https://github.com/composer/composer/issues/9084