:sparkles: New `Universal.CodeAnalysis.ForeachUniqueAssignment` sniff
Inspired by a tweet from dseguy, this new sniff will detect foreach control structures which use the same variable for both the _key_ as well as the _value_ assignment as this will lead to unexpected - and most likely unintended - behaviour.

The sniff includes a fixer which will remove the key assignment when the issue is detected as the value is required and the key optional.

The fixer maintains the existing behaviour for the code, though that behaviour may not be the desired behaviour, so any autogenerated fixes should be carefully reviewed (as always).

Refs:
* https://twitter.com/exakat/status/1509103728934203397
* https://3v4l.org/DdddX

Includes fixer.
Includes unit tests.
Includes documentation.