Composer: up the minimum PHPCS version to 3.7.0
PHPCS 3.7.0 has just been released and contains significant improvements to the - most notably - the tokenizer.

PHP 8.0 and 8.1 introduced a lot of new syntaxes and, while a lot of this can be supported for older PHPCS versions via PHPCSUtils, this will always impact performance and precision, so it is strongly advisable to use the latest PHPCS version as the new minimum to support the newly available PHP syntaxes.

Includes updating the GH Actions matrixes for this change.

Refs: https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.7.0