:sparkles: New `Universal.WhiteSpace.PrecisionAlignment` sniff
New sniff to enforce indentation to always be a multiple of a tabstop, i.e. disallow precision alignment.

For space-based standards, the `Generic.WhiteSpace.DisallowTabIndent` sniff (unintentionally/incorrectly) already removes precision alignment.
For tab-based standards, the `Generic.WhiteSpace.DisallowSpaceIndent` sniff, however, does not remove precision alignment.
With that in mind, this sniff is mostly intended for standards which demand tabs for indentation.

In rare cases, spaces for precision alignment can be intentional and acceptable, but more often than not, precision alignment is a typo.

Notes:
* This sniff does not concern itself with tabs versus spaces.
    Use the PHPCS native `Generic.WhiteSpace.DisallowTabIndent` or the `Generic.WhiteSpace.DisallowSpaceIndent` sniffs for that.
* When using this sniff with tab-based standards, please ensure that the `tab-width` is set and either don't set the `$indent` property or set it to the tab-width (or a multiple thereof).
* Precision alignment *within* text strings (multi-line text strings, heredocs, nowdocs) will NOT be flagged by this sniff.
* Precision alignment in (trailing) whitespace on otherwise blank lines will not be flagged by default.
    Most standards will automatically remove trailing whitespace anyway using the `Squiz.WhiteSpace.SuperfluousWhitespace` sniff.
    If the Squiz sniff is not used, set the `ignoreBlankLines` property to `false` to enable reporting for this.
* The sniff also supports an option to ignore precision alignment in specific situations, like for multi-line chained method calls.

**Note**: the fixer works based on "best guess" and may not always result in the desired indentation. Combine this sniff with the `Generic.WhiteSpace.ScopeIndent` sniff for more precise indentation fixes.
If so desired, the fixer can be turned off by including the rule in a standard like so: `<rule phpcs-only="true" ref="Universal.WhiteSpace.PrecisionAlignment"/>`

As it is, the sniff supports both PHP, as well as JS and CSS, though the support for JS and CSS should be considered incidental and will be removed once PHPCS 4.0 will be released.

Includes fixer.
Includes unit tests.
Includes documentation.