:sparkles: New `Universal.WhiteSpace.AnonClassKeywordSpacing` sniff
Checks the amount of spacing between the `class` keyword and the open parenthesis (if any) for anonymous class declarations.

The expected amount of spacing is configurable and defaults to `0`.

The default value of `0` has been decided upon based on scanning a not insignificant number of codebases and determining the prevailing style used for anonymous classes.
It is also in line with the previously proposed [errata for PSR12](https://github.com/php-fig/fig-standards/pull/1206).

Includes fixer.
Includes unit tests.
Includes documentation.
Includes metrics.