Universal/[Require/Disallow]FinalClass: rename metric
What with PHP 8.2 introducing `readonly` classes, the metric name `Class declaration type` would become ambiguous as `readonly` is not taken into account for the metric (and doesn't need to be as it doesn't have a direct relationship to `abstract`/`final`).

To prevent confusion about what the metric records, the metric name has been updated.