Universal/DisallowFinalClass: minor code tweak
When a class is declared as `abstract`, it would attempt to record both for the `abstract`, as well as the `not abstract, not final` metric.

As a metric can only be recorded once per token, this didn't actually cause any issues, but it still made the code confusing. Fixed now.