Universal/OneStatementInShortEchoTag: prevent false positive
In convoluted code, which I hope to never encounter in real life, it _could_ be possible to encounter a semi-colon, which is not the end of the statement, in the statement being executed for the `echo`.

This commit hardens the sniff to improve the handling of these.

Includes additional tests.