Universal/NoLeadingBackslash: examine imports within a group use statement
Previously the sniff would only examine the start of a complete import statement, now it will also examine the partial imports within a group `use` statement.

A partial import statement within a group use starting with a leading backslash is actually a parse error, which normally gets ignored by PHPCS, but this is a very specific one, which is also auto-fixable, so may as well report it.

This new error for leading backslashes for partial import statements within a group use statement will be reported using a separate error code `LeadingBackslashFoundInGroup`.

Includes updated and extra unit tests.