Universal/RequireFinalClass: always record the metric
As things were, the sniff would bow out if the opening brace for the class could not be found (yet) and would also not record the `not abstract, not final` metric in that case, even though for similar code using `abstract` or `final`, the metric _would_ be recorded.

This commit ensures the `not abstract, not final` metric will now be recorded either way.