ConstructorDestructorReturn: respect a potentially set `php_version` config value
PHP4-style constructors - methods with the same name as the class - are no longer regarded as constructors on PHP 8.0 or higher.

This adjusts the sniff to respect a potentially set `php_version` configuration value and will prevent the sniff from reporting on PHP4-style constructors when the `php_version` is set to PHP 8.0 or higher.

Includes unit tests.

Partially fixes 207