GH Actions: fix CI
Grrr....

When PHPUnit has been installed on a high PHP version, some of the dependencies of PHPUnit may now be installed in versions not compatible with PHP 7.4, which would block the install of the Coveralls package.

While this is not (yet) a problem in this repo as the max PHPUnit version is set to PHPUnit 7.x, the issue should still be prevented.

Installing PHP Coveralls globally instead should fix it.

I just wish PHP Coveralls would finally release a version compatible with PHP >  8.0....