:sparkles: New `Universal.UseStatements.DisallowMixedGroupUse` sniff
New sniff to disallow group use statements which combine imports for namespace/OO, functions and/or constants in one statement.

Note: the fixer will use a semi-standardized format for group use statements.
If there are more specific requirements for the formatting of group use statements, the ruleset configurator should ensure that additional sniffs are included in the ruleset to enforce the required format.

Includes unit tests.
Includes documentation.
Includes metrics.