Modernize/Dirname: respect a passed `php_version`
If no `php_version` is passed, the sniff will behave as before.

However, is a `php_version` is passed, the sniff will now selectively stay silent depending on the configured PHP and whether the proposed fixes could be applied.

Includes additional tests.
Includes adding information about the support of the `php_version` config setting to the README docs.