CodeAnalysis/ConstructorDestructorReturn: correctly ignore PHP-4-style constructors in namespaced classes
Methods with the same name as a class are not recognized as PHP4-style constructors in namespaced classes.

This was so far not taken into account in this sniff, leading to false positives.

A better solution, with a lower performance impact is expected at a later point in time (namespace tracker), however, as my time is limited, it may still be a while before that solution is available.

In the mean time, let's just fix it.

Includes test.