Universal/DuplicateArrayKey: add extra tests
Numeric strings which do not comply with the rules for PHP 7.4 numeric literals with underscores should not be regarded as numeric array indexes.
This bug was fixed in PHPCSUtils 1.1.0.

This commit adds some tests to safeguard the fix in the sniff for which this problem was originally reported.

See PHPCSStandards/PHPCSUtils 619