Universal/KeywordSpacing: update tests for PHPCS 4.0
Two tests from the `KeywordSpacingUnitTest.1.inc` file involve code which is interpreted differently in PHP < 8.0 vs 8.0 due to the PHP change to tokenize namespaced names as a single token.
Along the same lines, the code samples in the `KeywordSpacingUnitTest.2.inc` file will also no longer trigger the sniff.

As this change has been polyfilled in PHPCS 4.0, this invalidates those tests when running the sniff on PHPCS 4.0.

This commit moves the original tests from the `KeywordSpacingUnitTest.1.inc` file to a new `KeywordSpacingUnitTest.5.inc` file, which will be skipped on PHPCS 4.0. The `KeywordSpacingUnitTest.2.inc` file will also be skipped.
It also updates the tests in the `KeywordSpacingUnitTest.1.inc` file to be tests which are PHPCS cross-version compatible.