Universal/DisallowUseFunction: update for PHPCS 4.0
The token walking this sniff did to find the exact token to report the error on needed to be updated to allow for the PHP 8.0 namespaced name as single token change as applied in PHPCS 4.0.

Fixed now.

Includes updating one test to make it cross-version compatible as the original code was not valid since PHP 8.0.