Universal/NoUselessAliases: make variable name more descriptive
As of PHPCSUtils 1.0.12, the values in the return array from the `UseStatements::splitImportUseStatement()` method will now never include a leading backslash.

This commit improves the variable name used when walking the return value to make it abundantly clear that the "full name" in the return value from the method, is what PHP calls a "qualified name", i.e. without leading backslash.

Ref: PHPCSStandards/PHPCSUtils 590