Universal/CommaSpacing: fix for PHPCSUtils 1.1.0
As of PHPCS 4.x - and PHPCS cross-version as of PHPCSUtils 1.1.0, a closure `T_USE` token is now a parenthesis owner.

This also means that when PHPCSUtils 1.1.0 is used, the `T_USE` will be recognized as a parentheses owner, which means the dedicated error code is broken as the logic will go into the switch and end up using the default error code and will never reach the fall-back logic for "unowned" parentheses, in which this was previously handled.

The tests don't fail on this as the tests don't currently safeguard that the correct _error code_ is thrown, they just verify the number of errors and that was unchanged.

Either way, fixed now.

Includes minor inline doc clean-up.