Composer: update to PHPCSDevCS 1.2.0
PHPCSDevCS now allows for PHPCS 4.0 and includes PHPCompatibility 10.0.0-alpha1 and a range of sniffs from PHPCSExtra.

Includes preventing issues with recursive dependency.
Includes minor documentation update in the ruleset.
Includes renaming a parameter of a method overloading a PHPCS native method to be in-line with the upstream PHPCS 4.0 parameter name.
Includes various small CS fixes.

Includes running the CS check for this codebase against PHPCS 4.x-dev from now on.

Refs:
* https://github.com/PHPCSStandards/PHPCSDevCS/releases/tag/1.2.0