Composer: raise the minimum supported PHPCS + PHPCSUtils versions
PHP_CodeSniffer 3.13.3 + 3.13.4 and PHPCSUtils 1.1.2 have been released.
This commit raises the minimum supported PHPCS/Utils versions for improved PHP 8.4 support.

Note: while raising to 3.13.4 is not necessary from a functional point of view, 3.13.3 contains an issue which blocks the ability to run the tests, so we need 3.13.4 to safeguard the functionality of this external standard correctly.

Ref:
* https://github.com/PHPCSStandards/PHPCSUtils/releases/tag/1.1.2
* https://github.com/PHPCSStandards/PHP_CodeSniffer/releases/tag/3.13.3
* https://github.com/PHPCSStandards/PHP_CodeSniffer/releases/tag/3.13.4